requirements.txt
numpy>=1.26.0
pandas>=2.1.1
praw>=7.8.1
python-dotenv==1.0.0
requests>=2.31.0
yfinance>=0.2.36
textblob>=0.17.1